fix conflicts

+++ components/raftstore/src/store/peer_storage.rs
@@ -1225,7 +1225,7 @@ where
             let drain_to = if half { cache_len / 2 } else { cache_len - 1 };
             let idx = cache.cache[drain_to].index;
             let mem_size_change = cache.compact_to(idx + 1);
-            RAFT_ENTRIES_EVICT_BYTES.inc_by(mem_size_change);
+            RAFT_ENTRIES_EVICT_BYTES.inc_by(mem_size_change as i64);
         }
     }
